refactor : FlickrApp 가독성 향상을 위해 UI 분리
