+++ kwon-dae-won/app/src/main/java/kwon/dae/won/ui/FlickrApp.kt
@@ -1,124 +1,56 @@
 package kwon.dae.won.ui
 
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.material3.*
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TextField
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.paging.compose.LazyPagingItems
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.paging.compose.collectAsLazyPagingItems
 import coil.compose.AsyncImage
-import coil.compose.AsyncImagePainter
-import coil.compose.rememberAsyncImagePainter
-import kwon.dae.won.domain.model.Photo
-import timber.log.Timber
+import kwon.dae.won.FlickrViewModel
 
 
 /**
  * @author Daewon on 10,February,2023
  *
  */
-@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun FlickrApp(
-    photos: LazyPagingItems<Photo>,
-    onLongClick: (Int) -> Unit,
+    viewModel: FlickrViewModel = viewModel(),
 ) {
+    val lazyPagingPhotos = viewModel.getRecentPhotos().collectAsLazyPagingItems()
+    val photos by remember { mutableStateOf(lazyPagingPhotos) }
+    var openDialog by remember { mutableStateOf(false to -1) }
 
-    LazyVerticalGrid(
-        columns = GridCells.Adaptive(100.dp),
-    ) {
-        items(photos.itemCount) { index ->
-            Row(
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                if (photos[index]?.id?.isNotBlank() == true) {
-                    var isImageLoading by remember { mutableStateOf(false) }
+    Column(modifier = Modifier.fillMaxSize()) {
 
-                    val painter = rememberAsyncImagePainter(
-                        model = loadImageData(
-                            LocalContext.current,
-                            photos[index]?.id,
-                            photos[index]?.secret
-                        )
-                    )
+        CurrentPhotoList(photos = photos, onLongClick = { index ->
+            openDialog = true to index
+        })
+    }
 
-                    isImageLoading = when (painter.state) {
-                        is AsyncImagePainter.State.Loading -> true
-                        else -> false
-                    }
-
-                    Box(
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Image(
-                            modifier = Modifier
-                                .width(200.dp)
-                                .height(150.dp)
-                                .combinedClickable(
-                                    onClick = { },
-                                    onLongClick = { onLongClick(index) }
-                                ),
-                            painter = painter,
-                            contentDescription = "Photo Image",
-                            contentScale = ContentScale.FillBounds,
-                        )
-
-                        if (isImageLoading) {
-                            CircularProgressIndicator(
-                                modifier = Modifier
-                                    .padding(horizontal = 6.dp, vertical = 3.dp),
-                                color = MaterialTheme.colorScheme.primary
+    if (openDialog.first) {
+        photos[openDialog.second]?.let { item ->
+            DefaultAlertDialog(
+                url = imageUrl(item.id, item.secret),
+                onDismissButtonClick = { openDialog = false to -1 },
+                onConfirmButtonClick = {
+                    if (openDialog.second > 0) {
+                        photos[openDialog.second]?.let { item ->
+                            viewModel.downloadPhoto(
+                                item.title,
+                                item.owner,
+                                imageUrl(item.id, item.secret)
                             )
                         }
                     }
-
+                    openDialog = false to -1
                 }
-            }
+            )
         }
     }
-}
-
-@Composable
-fun DefaultAlertDialog(
-    url: String,
-    onConfirmButtonClick: () -> Unit,
-    onDismissButtonClick: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = { onDismissButtonClick() },
-        title = {
-            Text(text = "다운로드 하시겠습니까?")
-        },
-        text = {
-            Box(
-                modifier = Modifier.wrapContentSize().fillMaxWidth()
-            ) {
-                AsyncImage(modifier = Modifier.align(Alignment.Center),model = url, contentDescription = null)
-            }
-
-        },
-        dismissButton = {
-            TextButton(
-                onClick = { onDismissButtonClick() }
-            ) {
-                Text("취소")
-            }
-        },
-        confirmButton = {
-            TextButton(
-                onClick = { onConfirmButtonClick() }
-            ) {
-                Text("확인")
-            }
-        },
-    )
 }
